moved basic expansion algorithm to its own package
