--- expanders/expander.go
@@ -1,6 +1,0 @@
-package expanders
-
-// Expander defines the required behaviour for any token expander.
-type Expander interface {
-	Expand(token string) ([]string, error)
-}
